get realtime data from firestore
