+++ assignment2/Screens/AddActivity.js
@@ -33,21 +33,7 @@ const AddActivity = ({navigation}) => {
 
     const { activities, setActivities } = useActivities();
 
-    useEffect(()=>{
-      // set up a listener to get realtime data from firestore - only after the first render
-      onSnapshot(collection(database,"activities"),(querySnapshot)=>{
-        if (querySnapshot.empty){
-          Alert.alert("You need to add something");
-          return;
-        }
-        let newArray = []
-        // look through this querySnapshot
-        querySnapshot.forEach((doc) => {
-          newArray.push({...doc.data(),id:doc.id});
-      });
-      setActivities(newArray);
-      })
-    },[])
+    
 
     // validate if the user input is valid and send alerts to user
     function validateInput(){
